modulated azure storage api functions
